--- files/etc/indico/indico.conf
@@ -1,46 +1,0 @@
-import ast
-import os
-
-ATTACHMENT_STORAGE = os.environ.get('ATTACHMENT_STORAGE', 'default')
-AUTH_PROVIDERS = ast.literal_eval(os.environ.get('INDICO_AUTH_PROVIDERS', 'None')) or {}
-BASE_URL = "{SERVICE_SCHEME}://{SERVICE_HOSTNAME}".format(**os.environ)
-CACHE_DIR        = "/srv/indico/cache"
-CELERY_CONFIG = ast.literal_eval(os.environ.get('CELERY_CONFIG', 'None')) or {}
-CELERY_BROKER = os.environ['CELERY_BROKER']
-CUSTOMIZATION_DIR = '/srv/indico/custom'
-CUSTOMIZATION_DEBUG = os.environ.get('INDICO_CUSTOMIZATION_DEBUG', False)
-DEFAULT_LOCALE = os.environ.get('INDICO_DEFAULT_LOCALE', 'en_GB')
-DEFAULT_TIMEZONE = os.environ.get('INDICO_DEFAULT_TIMEZONE', 'UTC')
-ENABLE_ROOMBOOKING = os.environ.get('ENABLE_ROOMBOOKING', False)
-IDENTITY_PROVIDERS = ast.literal_eval(os.environ.get('INDICO_IDENTITY_PROVIDERS', 'None')) or {}
-LOCAL_IDENTITIES = not IDENTITY_PROVIDERS
-LOCAL_GROUPS = True
-LOG_DIR          = "/srv/indico/log"
-LOGGING_CONFIG_FILE = "/etc/logging.yaml"
-LOGO_URL = os.environ.get('INDICO_LOGO_URL')
-if os.path.isfile(CUSTOMIZATION_DIR + "/files/logo.png"):
-    LOGO_URL = "/static/custom/files/logo.png"
-NO_REPLY_EMAIL = os.environ.get('INDICO_NO_REPLY_EMAIL', "noreply-root@localhost")
-PLUGINS = tuple(x for x in os.environ.get('INDICO_EXTRA_PLUGINS', '').split(',') if x.strip())
-PROVIDER_MAP = ast.literal_eval(os.environ.get('INDICO_PROVIDER_MAP', 'None')) or {}
-PUBLIC_SUPPORT_EMAIL = os.environ.get('INDICO_PUBLIC_SUPPORT_EMAIL', "root@localhost")
-REDIS_CACHE_URL = os.environ['REDIS_CACHE_URL']
-ROUTE_OLD_URLS = os.environ.get('INDICO_ROUTE_OLD_URLS', False)
-SECRET_KEY = os.environ['SECRET_KEY']
-SENTRY_DSN = os.environ.get('SENTRY_DSN', None)
-if os.environ.get('SERVICE_PORT'):
-    BASE_URL += ':' + os.environ.get('SERVICE_PORT')
-SMTP_ALLOWED_SENDERS = {NO_REPLY_EMAIL}
-SMTP_LOGIN            = os.environ.get('SMTP_LOGIN')
-SMTP_SENDER_FALLBACK = NO_REPLY_EMAIL
-SMTP_SERVER           = os.environ.get('SMTP_SERVER'), os.environ.get('SMTP_PORT')
-SMTP_PASSWORD         = os.environ.get('SMTP_PASSWORD')
-SMTP_USE_TLS          = os.environ.get('SMTP_USE_TLS')
-STORAGE_BACKENDS = ast.literal_eval(os.environ.get('STORAGE_DICT'))
-SUPPORT_EMAIL = os.environ.get('INDICO_SUPPORT_EMAIL', "root@localhost")
-SYSTEM_NOTICES_URL = None
-SQLALCHEMY_DATABASE_URI = os.environ['INDICO_DB_URI']
-SQLALCHEMY_POOL_SIZE = 20
-SQLALCHEMY_POOL_TIMEOUT = 5
-TEMP_DIR         = "/srv/indico/tmp"
-USE_PROXY = True
