--- files/etc/indico/logging.yaml
@@ -1,71 +1,0 @@
-# This is the default indico logging configuration.
-# It uses the standard Python dictConfig syntax, but provides a few enhancements:
-# - The `filename` of a `logging.FileHandler` is relative to the configured LOG_DIR
-# - The SMTP connection data of a `logging.handlers.SMTPHandler` is taken from the
-#   indico config (this is disabled if a custom `mailhost` is set).
-#   `fromaddr`, `toaddrs` and `subject` are also set to useful default values, but
-#   these values may be overridden.
-# - The `indico.core.logger.FormattedSubjectSMTPHandler` handler allows formatting
-#   the email subject using a standard python format string (e.g. `%(message)s`).
-#   Besides that it behaves exactly like the normal `SMTPHandler`, including the
-#   Indico-specific default values mentioned above.  It also handles unicode
-#   properly so it is not recommended to use the default SMTPHandler as that one
-#   may silently drop some log messages which contain non-ascii characters.
-# - All log formatters may use `%(request_id)s` to get a random ID unique to each
-#   request which makes it easier to grep for all other log entries that belong to
-#   a specific request.
-# - All log formatters may also use `%(user_id)s` to get the ID of the user
-#   associated with the request context making it easier to grep for all log
-#   entries triggered by a specific user.
-# - Setting `append_request_info` to `true` in a formatter will append various
-#   information about the current request to the log message.  This will result in
-#   a multi-line message and is only recommended for errors; usually in a handler
-#   which sends emails.
-
-version: 1
-
-root:
-  level: INFO
-  handlers: [other]
-
-loggers:
-  indico:
-    handlers: [indico]
-  celery:
-    handlers: [celery, stderr]
-
-handlers:
-  indico:
-    class: logging.FileHandler
-    filename: indico.log
-    filters: [indico]
-    formatter: default
-  celery:
-    class: logging.FileHandler
-    filename: celery.log
-    filters: [celery]
-    formatter: simple
-  other:
-    class: logging.FileHandler
-    filename: other.log
-    filters: [other]
-    formatter: simple
-    level: WARNING
-  stderr:
-    class: logging.StreamHandler
-    formatter: default
-
-formatters:
-  default:
-    format: '%(asctime)s  %(levelname)-7s  %(request_id)s  %(user_id)-6s  %(name)-25s %(message)s'
-  simple:
-    format: '%(asctime)s  %(levelname)-7s  %(name)-25s %(message)s'
-
-filters:
-  indico:
-    name: indico
-  celery:
-    name: celery
-  other:
-    (): indico.core.logger.BlacklistFilter
-    names: [indico, celery]
