--- files/etc/nginx/nginx.conf
@@ -1,101 +1,0 @@
-user nginx nginx;
-daemon off;
-
-events {}
-http {
-  include mime.types;
-  server_tokens off;
-
-  add_header X-Content-Type-Options 'nosniff';
-  add_header X-Frame-Options 'SAMEORIGIN';
-
-  gzip on;
-  gzip_disable "msie6";
-  gzip_min_length 256;
-
-  gzip_proxied any;
-  gzip_http_version 1.1;
-  gzip_types
-   application/font-woff
-   application/font-woff2
-   application/x-javascript
-   application/xml
-   application/xml+rss
-   image/png
-   image/x-icon
-   font/woff2
-   text/css
-   text/javascript
-   text/plain
-   text/xml;
-
-  server {
-    listen 8080;
-    listen [::]:8080;
-    error_log stderr error;
-
-    location ~ ^/(images|fonts)(.*)/(.+?)(__v[0-9a-f]+)?\.([^.]+)$ {
-      alias /srv/indico/static/$1$2/$3.$5;
-    }
-
-    location ~ ^/(css|dist|images|fonts)/(.*)$ {
-      alias /srv/indico/static/$1/$2;
-      access_log off;
-    }
-
-    location /robots.txt {
-      alias /srv/indico/static/robots.txt;
-      access_log off;
-    }
-
-    location / {
-      root /var/empty/nginx;
-      proxy_hide_header Cache-Control;
-      add_header Cache-Control 'no-cache,private';
-      add_header X-Content-Type-Options 'nosniff';
-      add_header X-Frame-Options 'SAMEORIGIN';
-      client_max_body_size 0;
-      proxy_set_header X-Forwarded-Proto $http_x_forwarded_proto;
-      proxy_set_header X-Forwarded-Host $http_host;
-      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-      proxy_pass http://localhost:8081;
-
-      location ~ ^/static/custom/(.*)$ {
-        expires 1m;
-        proxy_hide_header Cache-Control;
-        add_header Cache-Control 'public';
-        add_header X-Content-Type-Options 'nosniff';
-        add_header X-Frame-Options 'SAMEORIGIN';
-        proxy_pass http://localhost:8081;
-      }
-      
-      location ~ ^/assets/(.*)$ {
-        proxy_hide_header Cache-Control;
-        add_header Cache-Control 'no-cache';
-        add_header X-Content-Type-Options 'nosniff';
-        add_header X-Frame-Options 'SAMEORIGIN';
-        proxy_pass http://localhost:8081;
-      }
-    }
-
-    location /health {
-      access_log off;
-      add_header 'Content-Type' 'application/json';
-      return 204;
-    }
-  }
-
-  server {
-    listen 9080;
-    listen [::]:9080;
-    error_log stderr error;
-
-    location / {
-      return 404;
-    }
-
-    location = /stub_status {
-      stub_status;
-    }
-  }
-}
