+++ indico-nginx.Dockerfile
@@ -2,9 +2,6 @@ FROM ubuntu:jammy as builder
 
 ENV DEBIAN_FRONTEND=noninteractive
 
-ARG nginx_gid=2001
-ARG nginx_uid=2001
-
 RUN apt-get update \
     && apt-get install -y --no-install-recommends \
         build-essential \
@@ -12,10 +9,8 @@ RUN apt-get update \
         python3-dev \
         python3-pip \
     && apt-get clean \
-    && rm -rf /var/lib/apt/lists/* \
-    && addgroup --gid ${nginx_gid} nginx \
-    && adduser --system --gid ${nginx_gid} --uid ${nginx_uid} --home /srv/indico --disabled-login nginx
-USER nginx
+    && rm -rf /var/lib/apt/lists/*
+
 RUN python3 -m pip install --no-cache-dir --no-warn-script-location --prefer-binary \
     indico==3.2.0 \
     indico-plugin-piwik \
@@ -28,8 +23,8 @@ RUN apt-get update \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
 
-COPY --from=builder /srv/indico/.local/lib/python3.10/site-packages/indico/web/static /srv/indico/static
-COPY files/etc/nginx/nginx.conf /etc/nginx/nginx.conf
+COPY --from=builder /usr/local/lib/python3.10/dist-packages/indico/web/static /srv/indico/static
+COPY nginx_rock/etc/nginx/nginx.conf /etc/nginx/nginx.conf
 
 ARG nginx_gid=2001
 ARG nginx_uid=2001
