+++ indico_rock/etc/indico/uwsgi.ini
@@ -1,0 +1,36 @@
+[uwsgi]
+uid = indico
+gid = indico
+umask = 027
+
+processes = 4
+enable-threads = true
+http-socket = 0.0.0.0:8081
+protocol = http
+
+master = true
+auto-procname = true
+procname-prefix-spaced = indico
+log-5xx = true
+disable-logging = true
+
+single-interpreter = true
+
+touch-reload = /srv/indico/indico.wsgi
+wsgi-file = /srv/indico/indico.wsgi
+
+ignore-sigpipe = true
+ignore-write-errors = true
+disable-write-exception = true
+
+vacuum = true
+memory-report = true
+max-requests = 2500
+harakiri = 900
+harakiri-verbose = true
+reload-on-rss = 2048
+evil-reload-on-rss = 8192
+
+offload-threads = 4
+
+stats-push = statsd:127.0.0.1:9125
