+++ indico_rock/plugins/autocreate/README.md
@@ -1,0 +1,3 @@
+# Autocreate Plugin
+
+Extends the indico CLI to add a non-interactive way to create users
