+++ indico_rock/plugins/autocreate/autocreate/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Create users non-interactively."""
