+++ indico_rock/plugins/autocreate/setup.py
@@ -1,0 +1,5 @@
+"""Extends the indico CLI to add a non-interactive way to create users."""
+
+from setuptools import setup
+
+setup()
