+++ indico_rock/rockcraft.yaml
@@ -1,0 +1,82 @@
+name: indico
+summary: Indico rock
+description: Indico OCI image for the Indico charm
+version: "1.0"
+base: bare
+build-base: ubuntu:22.04
+license: Apache-2.0
+env:
+  - LANG: C.UTF-8
+  - LC_ALL: C.UTF-8
+  - LC_LANG: C.UTF-8
+  - UWSGI_EMBED_PLUGINS: stats_pusher_statsd
+platforms:
+  amd64:
+parts:
+  add-user:
+    plugin: nil
+    overlay-script: |
+      addgroup --gid 2000 indico
+      adduser --system --gid 2000 --uid 2000 --home /srv/indico indico
+  indico:
+    after:
+      - add-user
+    build-packages:
+      - build-essential
+      - libldap2-dev
+      - libpq-dev
+      - libsasl2-dev
+      - libxmlsec1-dev
+      - pkg-config
+      - python3-dev
+    stage-packages:
+      - build-essential
+      - ca-certificates
+      - git-core
+      - lib6
+      - libglib2.0-data
+      - libldap2-dev
+      - libpq-dev
+      - libsasl2-dev
+      - libxmlsec1-dev
+      - pkg-config
+      - postgresql-client
+      - python3-dev
+      - python3-pip
+      - python3-venv
+      - shared-mime-info
+      - texlive-fonts-recommended
+      - texlive-plain-generic
+      - texlive-xetex
+    source: .
+    plugin: python
+    python-packages:
+      - indico==3.2.0
+      - indico-plugin-piwik
+      - indico-plugin-storage-s3
+      - python3-saml
+      - python-ldap
+      - uwsgi
+      - plugins/autocreate
+    override-prime: |
+      /bin/bash -c "mkdir -p --mode=775 srv/indico/{archive,cache,custom,etc,log,tmp}"
+      /bin/bash -c "chown indico:indico srv/indico srv/indico/{archive,cache,custom,etc,log,tmp}"
+      cp $CRAFT_PART_INSTALL/lib/python3.10/site-packages/indico/web/indico.wsgi srv/indico/indico.wsgi
+      cp -R $CRAFT_PART_INSTALL/lib/python3.10/site-packages/indico/web/static srv/indico/static
+    prime:
+      - srv/indico/*
+  copy-config:
+    plugin: dump
+    source: .
+    organize:
+      start-indico.sh: srv/indico/start-indico.sh
+    permissions:
+      - owner: 2000
+        group: 2000
+      - path: srv/indico/start-indico.sh
+        mode: 544
+      - path: etc/indico/
+        mode: 755
+    prime:
+      - etc/indico/*
+      - srv/indico/start-indico.sh
