+++ indico_rock/start-indico.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+set -e
+
+indico db prepare
+indico db upgrade
+indico db --all-plugins upgrade
+uwsgi --ini /etc/uwsgi.ini
