--- plugins/autocreate/README.md
@@ -1,3 +1,0 @@
-# Autocreate Plugin
-
-Extends the indico CLI to add a non-interactive way to create users
