--- plugins/autocreate/setup.cfg
@@ -1,28 +1,0 @@
-[metadata]
-name = indico-plugin-autocreate
-version = 3.2
-description = Extends the indico CLI to add a non-interactive way to create users
-long_description = file: README.md
-long_description_content_type = text/markdown; charset=UTF-8; variant=GFM
-url = https://github.com/canonical/indico-operator
-license = Apache License 2.0
-author = launchpad.net/~canonical-is-devops
-author_email = is-devops-team@canonical.com
-classifiers =
-    Environment :: Plugins
-    Environment :: Web Environment
-    License :: OSI Approved :: Apache Software License
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-python_requires = >=3.9.0, <3.11
-install_requires =
-    indico>=3.2
-
-[options.entry_points]
-indico.plugins =
-    autocreate = autocreate.plugin:AutocreatePlugin
