--- plugins/autocreate/setup.py
@@ -1,5 +1,0 @@
-"""Extends the indico CLI to add a non-interactive way to create users."""
-
-from setuptools import setup
-
-setup()
